Fix to #46
This adds an easy dir path check and creation if needed.  This is to cover a situation such as when Kitsu is being used with Prism.  Kitsu may add the Dept and Task to Prism's UI, but not actually make the directory structure until a scenefile is created.  This fix will create the path as needed.