Fixed #48 - Error when adding new Shortcut
It seems in the new version of Prism, the ProjectBrowser (SceneBrowser) refresh was causing an error.  It looks like it is because the new Prism is using threads for the refresh which doesn't play well with Resolve.

This updates the shortcut saving to use the main Prism PB refresh function which seems to work.  The only issue with it is that it will refresh the entire PB.